+++ src/Data/Aeson/Parser/Internal.hs
@@ -45,6 +45,8 @@ module Data.Aeson.Parser.Internal
     -- ** Handling objects with duplicate keys
     , fromListAccum
     , parseListNoDup
+    -- * Text literal unescaping
+    , unescapeText
     ) where
 
 import Prelude.Compat
@@ -71,27 +73,17 @@ import qualified Data.ByteString.Lazy.Char8 as C
 import qualified Data.ByteString.Builder as B
 import qualified Data.Scientific as Sci
 import Data.Aeson.Parser.Unescape (unescapeText)
+import Data.Aeson.Internal.Integer
 import Data.Aeson.Internal.Text
+import Data.Aeson.Internal.Word8
 
 -- $setup
 -- >>> :set -XOverloadedStrings
 -- >>> import Data.Aeson.Types
 
-#define BACKSLASH 92
-#define CLOSE_CURLY 125
-#define CLOSE_SQUARE 93
-#define COMMA 44
-#define DOUBLE_QUOTE 34
-#define OPEN_CURLY 123
-#define OPEN_SQUARE 91
-#define C_0 48
-#define C_9 57
-#define C_A 65
-#define C_F 70
-#define C_a 97
-#define C_f 102
-#define C_n 110
-#define C_t 116
+-------------------------------------------------------------------------------
+-- Parsers
+-------------------------------------------------------------------------------
 
 -- | Parse any JSON value.
 --
@@ -133,11 +125,11 @@ json' = value'
 -- mkObject outside of the recursive loop for proper inlining.
 
 object_ :: ([(Key, Value)] -> Either String Object) -> Parser Value -> Parser Value
-object_ mkObject val = {-# SCC "object_" #-} Object <$> objectValues mkObject key val
+object_ mkObject val = Object <$> objectValues mkObject key val
 {-# INLINE object_ #-}
 
 object_' :: ([(Key, Value)] -> Either String Object) -> Parser Value -> Parser Value
-object_' mkObject val' = {-# SCC "object_'" #-} do
+object_' mkObject val' = do
   !vals <- objectValues mkObject key' val'
   return (Object vals)
  where
@@ -151,7 +143,7 @@ objectValues :: ([(Key, Value)] -> Either String Object)
 objectValues mkObject str val = do
   skipSpace
   w <- A.peekWord8'
-  if w == CLOSE_CURLY
+  if w == W8_CLOSE_CURLY
     then A.anyWord8 >> return KM.empty
     else loop []
  where
@@ -162,9 +154,9 @@ objectValues mkObject str val = do
   loop acc = do
     k <- (str A.<?> "object key") <* skipSpace <* (char ':' A.<?> "':'")
     v <- (val A.<?> "object value") <* skipSpace
-    ch <- A.satisfy (\w -> w == COMMA || w == CLOSE_CURLY) A.<?> "',' or '}'"
+    ch <- A.satisfy (\w -> w == W8_COMMA || w == W8_CLOSE_CURLY) A.<?> "',' or '}'"
     let acc' = (k, v) : acc
-    if ch == COMMA
+    if ch == W8_COMMA
       then skipSpace >> loop acc'
       else case mkObject acc' of
         Left err -> fail err
@@ -172,11 +164,11 @@ objectValues mkObject str val = do
 {-# INLINE objectValues #-}
 
 array_ :: Parser Value -> Parser Value
-array_ val = {-# SCC "array_" #-} Array <$> arrayValues val
+array_ val = Array <$> arrayValues val
 {-# INLINE array_ #-}
 
 array_' :: Parser Value -> Parser Value
-array_' val = {-# SCC "array_'" #-} do
+array_' val = do
   !vals <- arrayValues val
   return (Array vals)
 {-# INLINE array_' #-}
@@ -185,14 +177,14 @@ arrayValues :: Parser Value -> Parser (Vector Value)
 arrayValues val = do
   skipSpace
   w <- A.peekWord8'
-  if w == CLOSE_SQUARE
+  if w == W8_CLOSE_SQUARE
     then A.anyWord8 >> return Vector.empty
     else loop [] 1
   where
     loop acc !len = do
       v <- (val A.<?> "json list value") <* skipSpace
-      ch <- A.satisfy (\w -> w == COMMA || w == CLOSE_SQUARE) A.<?> "',' or ']'"
-      if ch == COMMA
+      ch <- A.satisfy (\w -> w == W8_COMMA || w == W8_CLOSE_SQUARE) A.<?> "',' or ']'"
+      if ch == W8_COMMA
         then skipSpace >> loop (v:acc) (len+1)
         else return (Vector.reverse (Vector.fromListN len (v:acc)))
 {-# INLINE arrayValues #-}
@@ -209,13 +201,13 @@ value = jsonWith (pure . KM.fromList)
 --
 -- ==== __Examples__
 --
--- 'json' keeps only the first occurence of each key, using 'Data.Aeson.KeyMap.fromList'.
+-- 'json' keeps only the first occurrence of each key, using 'Data.Aeson.KeyMap.fromList'.
 --
 -- @
 -- 'json' = 'jsonWith' ('Right' '.' 'H.fromList')
 -- @
 --
--- 'jsonLast' keeps the last occurence of each key, using
+-- 'jsonLast' keeps the last occurrence of each key, using
 -- @'HashMap.Lazy.fromListWith' ('const' 'id')@.
 --
 -- @
@@ -239,18 +231,18 @@ jsonWith mkObject = fix $ \value_ -> do
   skipSpace
   w <- A.peekWord8'
   case w of
-    DOUBLE_QUOTE  -> A.anyWord8 *> (String <$> jstring_)
-    OPEN_CURLY    -> A.anyWord8 *> object_ mkObject value_
-    OPEN_SQUARE   -> A.anyWord8 *> array_ value_
-    C_f           -> string "false" $> Bool False
-    C_t           -> string "true" $> Bool True
-    C_n           -> string "null" $> Null
-    _              | w >= 48 && w <= 57 || w == 45
-                  -> Number <$> scientific
-      | otherwise -> fail "not a valid json value"
+    W8_DOUBLE_QUOTE  -> A.anyWord8 *> (String <$> jstring_)
+    W8_OPEN_CURLY    -> A.anyWord8 *> object_ mkObject value_
+    W8_OPEN_SQUARE   -> A.anyWord8 *> array_ value_
+    W8_f             -> string "false" $> Bool False
+    W8_t             -> string "true" $> Bool True
+    W8_n             -> string "null" $> Null
+    _                 | w >= W8_0 && w <= W8_9 || w == W8_MINUS
+                     -> Number <$> scientific
+      | otherwise    -> fail "not a valid json value"
 {-# INLINE jsonWith #-}
 
--- | Variant of 'json' which keeps only the last occurence of every key.
+-- | Variant of 'json' which keeps only the last occurrence of every key.
 jsonLast :: Parser Value
 jsonLast = jsonWith (Right . KM.fromListWith (const id))
 
@@ -291,22 +283,22 @@ jsonWith' mkObject = fix $ \value_ -> do
   skipSpace
   w <- A.peekWord8'
   case w of
-    DOUBLE_QUOTE  -> do
-                     !s <- A.anyWord8 *> jstring_
-                     return (String s)
-    OPEN_CURLY    -> A.anyWord8 *> object_' mkObject value_
-    OPEN_SQUARE   -> A.anyWord8 *> array_' value_
-    C_f           -> string "false" $> Bool False
-    C_t           -> string "true" $> Bool True
-    C_n           -> string "null" $> Null
-    _              | w >= 48 && w <= 57 || w == 45
-                  -> do
-                     !n <- scientific
-                     return (Number n)
-      | otherwise -> fail "not a valid json value"
+    W8_DOUBLE_QUOTE  -> do
+                       !s <- A.anyWord8 *> jstring_
+                       return (String s)
+    W8_OPEN_CURLY    -> A.anyWord8 *> object_' mkObject value_
+    W8_OPEN_SQUARE   -> A.anyWord8 *> array_' value_
+    W8_f             -> string "false" $> Bool False
+    W8_t             -> string "true" $> Bool True
+    W8_n             -> string "null" $> Null
+    _                 | w >= W8_0 && w <= W8_9 || w == W8_MINUS
+                     -> do
+                       !n <- scientific
+                       return (Number n)
+                      | otherwise -> fail "not a valid json value"
 {-# INLINE jsonWith' #-}
 
--- | Variant of 'json'' which keeps only the last occurence of every key.
+-- | Variant of 'json'' which keeps only the last occurrence of every key.
 jsonLast' :: Parser Value
 jsonLast' = jsonWith' (pure . KM.fromListWith (const id))
 
@@ -321,7 +313,7 @@ jsonNoDup' = jsonWith' parseListNoDup
 
 -- | Parse a quoted JSON string.
 jstring :: Parser Text
-jstring = A.word8 DOUBLE_QUOTE *> jstring_
+jstring = A.word8 W8_DOUBLE_QUOTE *> jstring_
 
 -- | Parse a JSON Key
 key :: Parser Key
@@ -331,32 +323,28 @@ key = Key.fromText <$> jstring
 jstring_ :: Parser Text
 {-# INLINE jstring_ #-}
 jstring_ = do
-  -- not sure whether >= or bit hackery is faster
-  -- perfectly, we shouldn't care, it's compiler job.
-  s <- A.takeWhile (\w -> w /= DOUBLE_QUOTE && w /= BACKSLASH && w >= 0x20 && w < 0x80)
-  let txt = unsafeDecodeASCII s
+  s <- A.takeWhile (\w -> w /= W8_DOUBLE_QUOTE && w /= W8_BACKSLASH && w >= 0x20 && w < 0x80)
   mw <- A.peekWord8
   case mw of
-    Nothing           -> fail "string without end"
-    Just DOUBLE_QUOTE -> A.anyWord8 $> txt
-    Just w | w < 0x20 -> fail "unescaped control character"
-    _                 -> jstringSlow s
+    Nothing              -> fail "string without end"
+    Just W8_DOUBLE_QUOTE -> A.anyWord8 $> unsafeDecodeASCII s
+    Just w | w < 0x20    -> fail "unescaped control character"
+    _                    -> jstringSlow s
 
 jstringSlow :: B.ByteString -> Parser Text
 {-# INLINE jstringSlow #-}
-jstringSlow s' = {-# SCC "jstringSlow" #-} do
+jstringSlow s' = do
   s <- A.scan startState go <* A.anyWord8
   case unescapeText (B.append s' s) of
     Right r  -> return r
     Left err -> fail $ show err
  where
-    startState              = False
+    startState                = False
     go a c
-      | a                  = Just False
-      | c == DOUBLE_QUOTE  = Nothing
-      | otherwise = let a' = c == backslash
+      | a                     = Just False
+      | c == W8_DOUBLE_QUOTE  = Nothing
+      | otherwise = let a' = c == W8_BACKSLASH
                     in Just a'
-      where backslash = BACKSLASH
 
 decodeWith :: Parser Value -> (Value -> Result a) -> L.ByteString -> Maybe a
 decodeWith p to s =
@@ -451,7 +439,7 @@ jsonEOF' = json' <* skipSpace <* endOfInput
 -- | The only valid whitespace in a JSON document is space, newline,
 -- carriage return, and tab.
 skipSpace :: Parser ()
-skipSpace = A.skipWhile $ \w -> w == 0x20 || w == 0x0a || w == 0x0d || w == 0x09
+skipSpace = A.skipWhile $ \w -> w == W8_SPACE || w == W8_NL || w == W8_CR || w == W8_TAB
 {-# INLINE skipSpace #-}
 
 ------------------ Copy-pasted and adapted from attoparsec ------------------
@@ -461,90 +449,34 @@ data SP = SP !Integer {-# UNPACK #-}!Int
 
 decimal0 :: Parser Integer
 decimal0 = do
-  let zero = 48
   digits <- A.takeWhile1 isDigit_w8
-  if B.length digits > 1 && B.unsafeHead digits == zero
+  if B.length digits > 1 && B.unsafeHead digits == W8_0
     then fail "leading zero"
     else return (bsToInteger digits)
 
 -- | Parse a JSON number.
 scientific :: Parser Scientific
 scientific = do
-  let minus = 45
-      plus  = 43
   sign <- A.peekWord8'
-  let !positive = sign == plus || sign /= minus
-  when (sign == plus || sign == minus) $
+  let !positive = not (sign == W8_MINUS)
+  when (sign == W8_PLUS || sign == W8_MINUS) $
     void A.anyWord8
 
   n <- decimal0
 
   let f fracDigits = SP (B.foldl' step n fracDigits)
                         (negate $ B.length fracDigits)
-      step a w = a * 10 + fromIntegral (w - 48)
+      step a w = a * 10 + fromIntegral (w - W8_0)
 
   dotty <- A.peekWord8
-  -- '.' -> ascii 46
   SP c e <- case dotty of
-              Just 46 -> A.anyWord8 *> (f <$> A.takeWhile1 isDigit_w8)
-              _       -> pure (SP n 0)
+              Just W8_DOT -> A.anyWord8 *> (f <$> A.takeWhile1 isDigit_w8)
+              _           -> pure (SP n 0)
 
   let !signedCoeff | positive  =  c
                    | otherwise = -c
 
-  let littleE = 101
-      bigE    = 69
-  (A.satisfy (\ex -> ex == littleE || ex == bigE) *>
+  (A.satisfy (\ex -> case ex of W8_e -> True; W8_E -> True; _ -> False) *>
       fmap (Sci.scientific signedCoeff . (e +)) (signed decimal)) <|>
     return (Sci.scientific signedCoeff    e)
 {-# INLINE scientific #-}
-
------------------- Copy-pasted and adapted from base ------------------------
-
-bsToInteger :: B.ByteString -> Integer
-bsToInteger bs
-    | l > 40    = valInteger 10 l [ fromIntegral (w - 48) | w <- B.unpack bs ]
-    | otherwise = bsToIntegerSimple bs
-  where
-    l = B.length bs
-
-bsToIntegerSimple :: B.ByteString -> Integer
-bsToIntegerSimple = B.foldl' step 0 where
-  step a b = a * 10 + fromIntegral (b - 48) -- 48 = '0'
-
--- A sub-quadratic algorithm for Integer. Pairs of adjacent radix b
--- digits are combined into a single radix b^2 digit. This process is
--- repeated until we are left with a single digit. This algorithm
--- performs well only on large inputs, so we use the simple algorithm
--- for smaller inputs.
-valInteger :: Integer -> Int -> [Integer] -> Integer
-valInteger = go
-  where
-    go :: Integer -> Int -> [Integer] -> Integer
-    go _ _ []  = 0
-    go _ _ [d] = d
-    go b l ds
-        | l > 40 = b' `seq` go b' l' (combine b ds')
-        | otherwise = valSimple b ds
-      where
-        -- ensure that we have an even number of digits
-        -- before we call combine:
-        ds' = if even l then ds else 0 : ds
-        b' = b * b
-        l' = (l + 1) `quot` 2
-
-    combine b (d1 : d2 : ds) = d `seq` (d : combine b ds)
-      where
-        d = d1 * b + d2
-    combine _ []  = []
-    combine _ [_] = errorWithoutStackTrace "this should not happen"
-
--- The following algorithm is only linear for types whose Num operations
--- are in constant time.
-valSimple :: Integer -> [Integer] -> Integer
-valSimple base = go 0
-  where
-    go r [] = r
-    go r (d : ds) = r' `seq` go r' ds
-      where
-        r' = r * base + fromIntegral d